fix(VCF URL): fix the default VCF file URLs for chr MT in Variation Pattern Finder and VCF to PED converter
Data for different chromsomes can come at a later date and does not follow LATEST_RELEASE_VCF template. In this case we set LATEST_RELEASE_VCF_$CHR to point to the specific file.

This change make sure we check for specific file for the chromosome first. If it is not set, then we follow the LATEST_RELEASE_VCF template.

Related to ENSEMBL-4516.